requirements.txt
#python -m venv StockAnal_Sys
#source ./StockAnal_Sys/bin/activate
# 基本面分析系统依赖包
# 核心数据处理
pandas>=2.0.0
numpy>=1.21.0
requests>=2.25.0
scipy>=1.7.0

# AI服务依赖
openai>=1.0.0
google-genai>=0.3.0

# 股票数据源
akshare>=1.11.0
yfinance>=0.2.18

# Web框架
flask>=2.0.0
flask-cors>=3.0.0
flask-caching>=2.0.0
flask-swagger-ui>=4.11.1

# 数据库支持
sqlalchemy>=1.4.0
sqlite3  # Python内置模块，无需安装

# 数据解析
beautifulsoup4>=4.9.1
lxml>=4.6.0
html5lib>=1.1

# 网络请求优化
urllib3>=1.26.0
certifi>=2021.5.25
charset-normalizer>=2.0.0

# 异步支持
aiohttp>=3.8.0
nest-asyncio>=1.5.0

# 数据处理工具
openpyxl>=3.0.9
xlrd>=2.0.1
tabulate>=0.8.9

# 进度条
tqdm>=4.62.0

# JavaScript引擎 (akshare需要)
mini-racer>=0.12.0

# 时间处理
python-dateutil>=2.8.2
pytz>=2021.3

# JSON处理
jsonpath>=0.82

# 额外数据源 (可选)
alpha-vantage>=2.3.1
finnhub-python>=2.4.0
quandl>=3.7.0

# 开发和测试工具
pytest>=6.0.0
pytest-cov>=2.12.0
black>=22.0.0
flake8>=4.0.0

# 日志和监控
loguru>=0.6.0
colorama>=0.4.4

# 缓存
redis>=4.0.0

# 配置管理
python-dotenv>=0.19.0
PyYAML>=6.0.0

# 图表和可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 数据验证
pydantic>=1.8.0

# HTTP客户端增强
httpx>=0.24.0

# 调试工具
debugpy>=1.6.0
ipython>=8.0.0

# 安全相关
cryptography>=3.4.8

# 机器学习
scikit-learn>=1.0.0
statsmodels>=0.13.0

# 部署相关
gunicorn>=20.0.0
supervisor>=4.2.0

# AI搜索工具
tavily-python>=0.3.0

# Google API客户端
google>=3.0.0